+++ packages/react-native/tsup.config.ts
@@ -1,11 +1,17 @@
-import { defineConfig } from "tsup"
+import { defineConfig } from "tsup";
 
 export default defineConfig({
   entry: ["src/index.ts"],
-  format: ["esm"],
+  format: ["esm", "cjs"],
   dts: true,
   clean: true,
   minify: false,
   sourcemap: true,
-  external: ["react", "react-native", "@factorialco/factorial-one-core"],
-})
+  external: ["react", "react-native"],
+  esbuildOptions(options) {
+    options.jsx = "automatic";
+    return options;
+  },
+  treeshake: true,
+  splitting: false,
+});